Creates a default watcher configuration file

Creates a default watcher configuration file

Usage

generate_nmm_config(
  .mod,
  model_number = "",
  files_to_track = c("lst", "ext", "grd"),
  tmp_dir = "/tmp",
  watched_dir = file.path("model", "nonmem"),
  output_dir = file.path(watched_dir, "in_progress"),
  poll_duration = 1,
  alert = "None",
  level = "Debug",
  email = "",
  threads = 1,
  topic = ""
)

Arguments

.mod

a bbi nonmem model object

model_number

a string of model number e.g. 101a15. Default is pulled from .mod

files_to_track

a vector of file extensions. Default is c("lst", "ext", "grd")

tmp_dir

a temporary directory location to run nonmem. Default is /tmp

watched_dir

the directory where nonmem will output. Default is here::here()/model/nonmem

output_dir

the directory where watcher will place log and any output files default is here::here()/model/nonmem/in_progress

poll_duration

the amount of time in seconds between the watchers polling

alert

Where to send alerts if any. Default is None, available options are None or Slack

level

What level to log at. Default is info. Available options are Trace, Debug, Info, Warn, Fatal

email

if alert is set to Slack, this should be the email associated with slack to get messages sent directly to you.

threads

number of threads if running a parallel job. Default is 1

topic

the ntfy.sh topic to send alerts to.

Value

none

Examples

if (FALSE) { # \dontrun{
  generate_nmm_config(.mod)
} # }