get partition suggestions

In a call to submit_slurm_model(), if the number of requested CPUs exceeds the number of CPUs available in the requested partition, check_slurm_partitions() errors.
This function follows up with a message providing one or two suggestions for the partition with the smallest sufficient number of CPUs and least amount of memory.
If there are no partitions with enough CPUs to accommodate the number requested, this function's return message clarifies this.

Usage

partition_advice(ncpu, partition, avail_cpus_table, cache)

Arguments

ncpu

number of CPUs requested by user

partition

name of partition requested by user

avail_cpus_table

table of partitions with respective number of CPUs and memory

cache

optional argument to forgo caching

Value

string with suggestion upon check_slurm_partitions() error