slurmtools: An R package for easily submitting NONMEM jobs to slurm.

This package aims to ease the submission and monitoring of NONMEM jobs running on slurm.

submitting jobs

submit_slurm_job: Submits a job to slurm

generate_nmm_config: Generates a NONMEMmonitor config file

monitoring jobs

get_slurm_jobs: Gives a table of submitted jobs to slurm that shows status and other information given with squeue

cancelling jobs

cancel_slurm_job: Cancels the specified job

slurm partitions

get_slurm_partitions: Gives a vector of available partitions a user can submit jobs to.

Author

Maintainer: Devin Pastoor devin@a2-ai.com

Authors:

Jenna Elwing jenna@a2-ai.com

Matthew Smith matthews@a2-ai.com