submit a nonmem model to slurm in parallel

Usage

submit_nonmem_model(
  .mod,
  partition = get_slurm_partitions(),
  ncpu = 1,
  overwrite = FALSE,
  dry_run = FALSE,
  ...,
  slurm_job_template_path = getOption("slurmtools.slurm_job_template_path"),
  submission_root = getOption("slurmtools.submission_root"),
  bbi_config_path = getOption("slurmtools.bbi_config_path"),
  slurm_template_opts = list()
)

Arguments

.mod

a path to a model or a bbi nonmem model object

partition

name of the partition to submit the model

ncpu

number of cpus to run the model against

overwrite

whether to overwrite existing model results

dry_run

return the command that would have been invoked, without invoking

...

arguments to pass to processx::run

slurm_job_template_path

path to slurm job template

submission_root

directory to track job submission scripts and output

bbi_config_path

path to bbi.yaml file for bbi configuration

slurm_template_opts

choose slurm template