Creates a default watcher configuration file

Creates a default watcher configuration file

Usage

generate_nmm_config(
  .mod,
  model_number = NULL,
  files_to_track = NULL,
  tmp_dir = NULL,
  watched_dir = file.path("model", "nonmem"),
  output_dir = file.path(watched_dir, "in_progress"),
  poll_duration = NULL,
  level = NULL,
  threads = NULL,
  alerter_opts = list()
)

Arguments

.mod

a bbi nonmem model object

model_number

a string of model number e.g. 101a15. Default is pulled from .mod

files_to_track

a vector of file extensions. If left blank, nmm will use .ext, .lst. and .grd.

tmp_dir

a temporary directory location to run nonmem. If left blank nmm will use /tmp

watched_dir

the directory where nonmem will output. Default is here::here()/model/nonmem

output_dir

the directory where watcher will place log and any output files default is here::here()/model/nonmem/in_progress

poll_duration

the amount of time in seconds between the watchers polling. If left blank, nmm will use 1 second

level

What level to log at. Available options are Trace, Debug, Info, Warn, Fatal if left blank, nmm will use Info

threads

number of threads if running a parallel job. If left blank nmm will use 1

alerter_opts

a list for setting up an alerter fields needed are alerter, command, message_flag, use_stdout, args (list of additional args to be passed to alerter)

Value

none

Examples

if (FALSE) { # \dontrun{
  generate_nmm_config(.mod)
} # }